+++ src/Diag/Result.hs
@@ -30,6 +30,7 @@ module Diag.Result (
   -- * Rendering
   renderFailure,
   renderSuccess,
+  renderFailureWithoutWarnings,
 ) where
 
 import Data.List.NonEmpty (NonEmpty)
@@ -93,6 +94,10 @@ data ErrWithStack = ErrWithStack Stack SomeErr
 
 -- | A stacktrace
 newtype Stack = Stack [Text]
+  deriving (Show)
+
+-- | Dictates whether or not we want to show traceback
+data TracebackStyle = None | Default
   deriving (Show)
 
 instance Functor Result where
@@ -160,18 +165,10 @@ resultToMaybe (Failure _ _) = Nothing
 --
 -- renderFailure displays all types of emitted warnings.
 renderFailure :: [EmittedWarn] -> ErrGroup -> Doc AnsiStyle -> Doc AnsiStyle
-renderFailure ws (ErrGroup _ ectx es) headerDoc = header headerDoc <> renderedCtx <> renderedErrs <> renderedPossibleErrs
+renderFailure ws (ErrGroup _ ectx es) headerDoc = header headerDoc <> renderErrCtxs ectx <> renderedErrsWithAnnotation <> renderedPossibleErrs
   where
-    renderedCtx :: Doc AnsiStyle
-    renderedCtx =
-      case ectx of
-        [] -> emptyDoc
-        _ -> section "Details" (vsep (map (\ctx -> renderErrCtx ctx <> line) ectx))
-
-    renderedErrs :: Doc AnsiStyle
-    renderedErrs =
-      section "Relevant errors" $
-        subsection "Error" (map renderErrWithStack (NE.toList es))
+    renderedErrsWithAnnotation :: Doc AnsiStyle
+    renderedErrsWithAnnotation = section "Relevant errors" $ subsection "Error" $ renderErrs es Default
 
     renderedPossibleErrs :: Doc AnsiStyle
     renderedPossibleErrs =
@@ -179,7 +176,10 @@ renderFailure ws (ErrGroup _ ectx es) headerDoc = header headerDoc <> renderedCt
         [] -> emptyDoc
         _ ->
           section "Possibly-related warnings" $
-            subsection "Warning" (map renderEmittedWarn ws)
+            unannotatedSubsection (map (renderEmittedWarn Default) ws)
+
+renderFailureWithoutWarnings :: ErrGroup -> Doc AnsiStyle -> Doc AnsiStyle
+renderFailureWithoutWarnings (ErrGroup _ ectx es) headerDoc = header headerDoc <> renderErrCtxs ectx <> section "Relevant errors" (subsection "Error" $ renderErrs es None)
 
 -- | renderSuccess turns a list of warnings from a Success into a message
 -- suitable for logging
@@ -195,7 +195,7 @@ renderSuccess :: [EmittedWarn] -> Doc AnsiStyle -> Maybe (Doc AnsiStyle)
 renderSuccess ws headerDoc =
   case notIgnoredErrs of
     [] -> Nothing
-    ws' -> Just $ header headerDoc <> subsection "Warning" (map renderEmittedWarn ws')
+    ws' -> Just $ header headerDoc <> unannotatedSubsection (map (renderEmittedWarn Default) ws')
   where
     notIgnoredErrs :: [EmittedWarn]
     notIgnoredErrs = filter (not . isIgnoredErrGroup) ws
@@ -204,24 +204,38 @@ renderSuccess ws headerDoc =
     isIgnoredErrGroup IgnoredErrGroup{} = True
     isIgnoredErrGroup _ = False
 
----------- Renering individual Result components: ErrCtx, EmittedWarn, SomeWarn, ErrWithStack
+---------- Rendering a collection of Result components: [ErrCtx], NonEmpty ErrWithStack
+
+renderErrCtxs :: [ErrCtx] -> Doc AnsiStyle
+renderErrCtxs errCtxs = case errCtxs of
+  [] -> emptyDoc
+  _ -> section "Details" (vsep (map (\ctx -> renderErrCtx ctx <> line) errCtxs))
+
+renderErrs :: NonEmpty ErrWithStack -> TracebackStyle -> [Doc AnsiStyle]
+renderErrs es tracebackStyle = map (`renderErrWithStack` tracebackStyle) (NE.toList es)
+
+---------- Rendering individual Result components: ErrCtx, EmittedWarn, SomeWarn, ErrWithStack
 
 renderErrCtx :: ErrCtx -> Doc AnsiStyle
 renderErrCtx (ErrCtx ctx) = renderDiagnostic ctx
 
-renderErrWithStack :: ErrWithStack -> Doc AnsiStyle
-renderErrWithStack (ErrWithStack (Stack stack) (SomeErr err)) =
-  renderDiagnostic err
-    <> line
-    <> line
-    <> annotate (color Cyan) "Traceback:"
-    <> line
-    <> case stack of
-      [] -> indent 2 "(none)"
-      _ -> indent 2 (vsep (map (pretty . ("- " <>)) stack))
+renderErrWithStack :: ErrWithStack -> TracebackStyle -> Doc AnsiStyle
+renderErrWithStack (ErrWithStack (Stack stack) (SomeErr err)) tracebackStyle = do
+  let traceback = case tracebackStyle of
+        Default ->
+          line
+            <> line
+            <> annotate (color Cyan) "Traceback:"
+            <> line
+            <> case stack of
+              [] -> indent 2 "(none)"
+              _ -> indent 2 (vsep (map (pretty . ("- " <>)) stack))
+        None -> ""
 
-renderEmittedWarn :: EmittedWarn -> Doc AnsiStyle
-renderEmittedWarn (IgnoredErrGroup ectx es) = renderedCtx <> renderedErrors
+  renderDiagnostic err <> traceback
+
+renderEmittedWarn :: TracebackStyle -> EmittedWarn -> Doc AnsiStyle
+renderEmittedWarn tracebackStyle (IgnoredErrGroup ectx es) = renderedCtx <> renderedErrors
   where
     renderedCtx =
       case ectx of
@@ -229,12 +243,9 @@ renderEmittedWarn (IgnoredErrGroup ectx es) = renderedCtx <> renderedErrors
         _ ->
           (vsep (map (\ctx -> renderErrCtx ctx <> line) ectx))
 
-    renderedErrors =
-      section
-        "Relevant errors"
-        $ subsection "Error" (map renderErrWithStack (NE.toList es))
-renderEmittedWarn (StandaloneWarn (SomeWarn warn)) = renderDiagnostic warn
-renderEmittedWarn (WarnOnErrGroup ws ectx es) = renderedWarnings <> renderedCtx <> renderedErrors
+    renderedErrors = subsection "Warning" (map (`renderErrWithStack` tracebackStyle) (NE.toList es))
+renderEmittedWarn _ (StandaloneWarn (SomeWarn warn)) = renderDiagnostic warn
+renderEmittedWarn tracebackStyle (WarnOnErrGroup ws ectx es) = renderedWarnings <> renderedCtx <> renderedErrors
   where
     renderedWarnings = vsep (map (\w -> renderSomeWarn w <> line) (NE.toList ws)) <> line
 
@@ -246,10 +257,7 @@ renderEmittedWarn (WarnOnErrGroup ws ectx es) = renderedWarnings <> renderedCtx 
             "Details"
             (vsep (map (\ctx -> renderErrCtx ctx <> line) ectx))
 
-    renderedErrors =
-      section
-        "Relevant errors"
-        $ subsection "Error" (map renderErrWithStack (NE.toList es))
+    renderedErrors = subsection "Warning" (map (`renderErrWithStack` tracebackStyle) (NE.toList es))
 
 renderSomeWarn :: SomeWarn -> Doc AnsiStyle
 renderSomeWarn (SomeWarn w) = renderDiagnostic w
@@ -274,3 +282,6 @@ section name content =
 
 subsection :: Doc AnsiStyle -> [Doc AnsiStyle] -> Doc AnsiStyle
 subsection name = vsep . map (\single -> annotate (color Yellow) name <> line <> line <> indent 2 single <> line)
+
+unannotatedSubsection :: [Doc AnsiStyle] -> Doc AnsiStyle
+unannotatedSubsection = vsep . map (\single -> indent 2 single <> line)