+++ src/Strategy/Glide.hs
@@ -2,7 +2,7 @@ module Strategy.Glide (
   discover,
 ) where
 
-import App.Fossa.Analyze.Types (AnalyzeProject (analyzeProject'), analyzeProject)
+import App.Fossa.Analyze.Types (AnalyzeProject (analyzeProjectStaticOnly), analyzeProject)
 import Control.Effect.Diagnostics (Diagnostics, context)
 import Control.Effect.Reader (Reader)
 import Data.Aeson (ToJSON)
@@ -42,7 +42,7 @@ instance ToJSON GlideProject
 
 instance AnalyzeProject GlideProject where
   analyzeProject _ = getDeps
-  analyzeProject' _ = getDeps
+  analyzeProjectStaticOnly _ = getDeps
 
 mkProject :: GlideProject -> DiscoveredProject GlideProject
 mkProject project =